requirements.txt
flask==2.3.2
gunicorn==21.2.0
pymongo[srv]==4.6.1
python-dotenv==1.0.0
